# SUS-4084: Optimize Forum Activity module query