--- a/extensions/wikia/Forum/Forum.setup.php
+++ b/extensions/wikia/Forum/Forum.setup.php
@@ -7,7 +7,7 @@
  *
  */
 
-$wgExtensionCredits['specialpage'][] = [
+$GLOBALS['wgExtensionCredits']['specialpage'][] = [
 	'name' => 'Forum',
 	'author' => [ 'Hyun Lim', 'Kyle Florence', 'Saipetch Kongkatong', 'Tomasz Odrobny' ],
 	'url' => 'https://github.com/Wikia/app/tree/dev/extensions/wikia/Forum',
@@ -17,78 +17,82 @@ $wgExtensionCredits['specialpage'][] = [
 $dir = __DIR__ . '/';
 
 // classes
-$wgAutoloadClasses['ForumSpecialController'] =  $dir . 'ForumSpecialController.class.php' ;
-$wgAutoloadClasses['ForumHooksHelper'] =  $dir . 'ForumHooksHelper.class.php' ;
-$wgAutoloadClasses['ForumController'] =  $dir . 'ForumController.class.php' ;
-$wgAutoloadClasses['ForumNotificationPlugin'] =  $dir . 'ForumNotificationPlugin.class.php' ;
-$wgAutoloadClasses['Forum'] =  $dir . 'Forum.class.php' ;
-$wgAutoloadClasses['ForumBoard'] =  $dir . 'ForumBoard.class.php' ;
-$wgAutoloadClasses['ForumBoardInfo'] =  $dir . 'ForumBoardInfo.class.php' ;
-$wgAutoloadClasses['ForumPostInfo'] =  $dir . 'ForumPostInfo.class.php' ;
-$wgAutoloadClasses['ForumHelper'] =  $dir . 'ForumHelper.class.php' ;
-$wgAutoloadClasses['ForumExternalController'] =  $dir . 'ForumExternalController.class.php' ;
-$wgAutoloadClasses['RelatedForumDiscussionController'] =  $dir . 'RelatedForumDiscussionController.class.php' ;
-$wgAutoloadClasses['ThreadWatchlistDeleteUpdate'] = $dir . 'ThreadWatchlistDeleteUpdate.php';
+$GLOBALS['wgAutoloadClasses']['ForumSpecialController'] =  $dir . 'ForumSpecialController.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumHooksHelper'] =  $dir . 'ForumHooksHelper.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumController'] =  $dir . 'ForumController.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumNotificationPlugin'] =  $dir . 'ForumNotificationPlugin.class.php' ;
+$GLOBALS['wgAutoloadClasses']['Forum'] =  $dir . 'Forum.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumBoard'] =  $dir . 'ForumBoard.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumBoardInfo'] =  $dir . 'ForumBoardInfo.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumPostInfo'] =  $dir . 'ForumPostInfo.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumHelper'] =  $dir . 'ForumHelper.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ForumExternalController'] =  $dir . 'ForumExternalController.class.php' ;
+$GLOBALS['wgAutoloadClasses']['RelatedForumDiscussionController'] =  $dir . 'RelatedForumDiscussionController.class.php' ;
+$GLOBALS['wgAutoloadClasses']['ThreadWatchlistDeleteUpdate'] = $dir . 'ThreadWatchlistDeleteUpdate.php';
+
+$GLOBALS['wgAutoloadClasses']['ForumActivityService'] = __DIR__ . '/ForumActivityService.php';
+$GLOBALS['wgAutoloadClasses']['CachedForumActivityService'] = __DIR__ . '/CachedForumActivityService.php';
+$GLOBALS['wgAutoloadClasses']['DatabaseForumActivityService'] = __DIR__ . '/DatabaseForumActivityService.php';
 
 // i18n mapping
-$wgExtensionMessagesFiles['Forum'] = $dir . 'Forum.i18n.php' ;
-$wgExtensionMessagesFiles['ForumAliases'] = $dir . 'Forum.alias.php';
+$GLOBALS['wgExtensionMessagesFiles']['Forum'] = $dir . 'Forum.i18n.php' ;
+$GLOBALS['wgExtensionMessagesFiles']['ForumAliases'] = $dir . 'Forum.alias.php';
 
 // special pages
-$wgSpecialPages['Forum'] =  'ForumSpecialController';
+$GLOBALS['wgSpecialPages']['Forum'] =  'ForumSpecialController';
 
 // hooks
-$wgHooks['AfterWallWikiActivityFilter'][] = 'ForumHooksHelper::onAfterWallWikiActivityFilter';
-$wgHooks['WallContributionsLine'][] = 'ForumHooksHelper::onWallContributionsLine';
-$wgHooks['getUserPermissionsErrors'][] = 'ForumHooksHelper::getUserPermissionsErrors';
-$wgHooks['WallRecentchangesMessagePrefix'][] = 'ForumHooksHelper::onWallRecentchangesMessagePrefix';
-$wgHooks['WallThreadHeader'][] = 'ForumHooksHelper::onWallThreadHeader';
-$wgHooks['WallMessageGetWallOwnerName'][] = 'ForumHooksHelper::onWallMessageGetWallOwnerName';
+$GLOBALS['wgHooks']['AfterWallWikiActivityFilter'][] = 'ForumHooksHelper::onAfterWallWikiActivityFilter';
+$GLOBALS['wgHooks']['WallContributionsLine'][] = 'ForumHooksHelper::onWallContributionsLine';
+$GLOBALS['wgHooks']['getUserPermissionsErrors'][] = 'ForumHooksHelper::getUserPermissionsErrors';
+$GLOBALS['wgHooks']['WallRecentchangesMessagePrefix'][] = 'ForumHooksHelper::onWallRecentchangesMessagePrefix';
+$GLOBALS['wgHooks']['WallThreadHeader'][] = 'ForumHooksHelper::onWallThreadHeader';
+$GLOBALS['wgHooks']['WallMessageGetWallOwnerName'][] = 'ForumHooksHelper::onWallMessageGetWallOwnerName';
 
-$wgHooks['WallHistoryThreadHeader'][] = 'ForumHooksHelper::onWallHistoryThreadHeader';
-$wgHooks['WallHistoryHeader'][] = 'ForumHooksHelper::onWallHistoryHeader';
+$GLOBALS['wgHooks']['WallHistoryThreadHeader'][] = 'ForumHooksHelper::onWallHistoryThreadHeader';
+$GLOBALS['wgHooks']['WallHistoryHeader'][] = 'ForumHooksHelper::onWallHistoryHeader';
 
-$wgHooks['WallHeader'][] = 'ForumHooksHelper::onWallHeader';
-$wgHooks['WallNewMessage'][] = 'ForumHooksHelper::onWallNewMessage';
-$wgHooks['WallBeforeRenderThread'][] = 'ForumHooksHelper::onWallBeforeRenderThread';
-$wgHooks['AfterBuildNewMessageAndPost'][] = 'ForumHooksHelper::onAfterBuildNewMessageAndPost';
-$wgHooks['WallMessageDeleted'][] = 'ForumHooksHelper::onWallMessageDeleted';
-$wgHooks['ContributionsLineEnding'][] = 'ForumHooksHelper::onContributionsLineEnding';
-$wgHooks['OasisAddPageDeletedConfirmationMessage'][] = 'ForumHooksHelper::onOasisAddPageDeletedConfirmationMessage';
-$wgHooks['FilePageImageUsageSingleLink'][] = 'ForumHooksHelper::onFilePageImageUsageSingleLink';
-$wgHooks['AfterPageHeaderPageSubtitle'][] = 'ForumHooksHelper::onAfterPageHeaderPageSubtitle';
-$wgHooks['PageHeaderActionButtonShouldDisplay'][] = 'ForumHooksHelper::onPageHeaderActionButtonShouldDisplay';
+$GLOBALS['wgHooks']['WallHeader'][] = 'ForumHooksHelper::onWallHeader';
+$GLOBALS['wgHooks']['WallNewMessage'][] = 'ForumHooksHelper::onWallNewMessage';
+$GLOBALS['wgHooks']['WallBeforeRenderThread'][] = 'ForumHooksHelper::onWallBeforeRenderThread';
+$GLOBALS['wgHooks']['AfterBuildNewMessageAndPost'][] = 'ForumHooksHelper::onAfterBuildNewMessageAndPost';
+$GLOBALS['wgHooks']['WallMessageDeleted'][] = 'ForumHooksHelper::onWallMessageDeleted';
+$GLOBALS['wgHooks']['ContributionsLineEnding'][] = 'ForumHooksHelper::onContributionsLineEnding';
+$GLOBALS['wgHooks']['OasisAddPageDeletedConfirmationMessage'][] = 'ForumHooksHelper::onOasisAddPageDeletedConfirmationMessage';
+$GLOBALS['wgHooks']['FilePageImageUsageSingleLink'][] = 'ForumHooksHelper::onFilePageImageUsageSingleLink';
+$GLOBALS['wgHooks']['AfterPageHeaderPageSubtitle'][] = 'ForumHooksHelper::onAfterPageHeaderPageSubtitle';
+$GLOBALS['wgHooks']['PageHeaderActionButtonShouldDisplay'][] = 'ForumHooksHelper::onPageHeaderActionButtonShouldDisplay';
 
 // notification hooks
-$wgHooks['NotificationGetNotificationMessage'][] = 'ForumNotificationPlugin::onGetNotificationMessage';
+$GLOBALS['wgHooks']['NotificationGetNotificationMessage'][] = 'ForumNotificationPlugin::onGetNotificationMessage';
 
 // forum discussion on article
 // It need to be first one !!!
-array_splice( $wgHooks['OutputPageBeforeHTML'], 0, 0, 'ForumHooksHelper::onOutputPageBeforeHTML' );
+array_splice( $GLOBALS['wgHooks']['OutputPageBeforeHTML'], 0, 0, 'ForumHooksHelper::onOutputPageBeforeHTML' );
 
-$wgHooks['WallAction'][] = 'ForumHooksHelper::onWallAction';
-$wgHooks['WallBeforeStoreRelatedTopicsInDB'][] = 'ForumHooksHelper::onWallStoreRelatedTopicsInDB';
-$wgHooks['WallAfterStoreRelatedTopicsInDB'][] = 'ForumHooksHelper::onWallStoreRelatedTopicsInDB';
+$GLOBALS['wgHooks']['WallAction'][] = 'ForumHooksHelper::onWallAction';
+$GLOBALS['wgHooks']['WallBeforeStoreRelatedTopicsInDB'][] = 'ForumHooksHelper::onWallStoreRelatedTopicsInDB';
+$GLOBALS['wgHooks']['WallAfterStoreRelatedTopicsInDB'][] = 'ForumHooksHelper::onWallStoreRelatedTopicsInDB';
 
-$wgHooks['ArticleFromTitle'][] = 'ForumHooksHelper::onArticleFromTitle';
+$GLOBALS['wgHooks']['ArticleFromTitle'][] = 'ForumHooksHelper::onArticleFromTitle';
 
 // For activity module tag
-$wgHooks['ParserFirstCallInit'][] = 'ForumHooksHelper::onParserFirstCallInit';
+$GLOBALS['wgHooks']['ParserFirstCallInit'][] = 'ForumHooksHelper::onParserFirstCallInit';
 
 // Hook for topic red links
-$wgHooks['LinkBegin'][] = 'ForumHooksHelper::onLinkBegin';
+$GLOBALS['wgHooks']['LinkBegin'][] = 'ForumHooksHelper::onLinkBegin';
 
 // Fix URLs of thread pages when purging them.
-$wgHooks['TitleGetSquidURLs'][] = 'ForumHooksHelper::onTitleGetSquidURLs';
-$wgHooks['ArticleCommentGetSquidURLs'][] = 'ForumHooksHelper::onArticleCommentGetSquidURLs';
+$GLOBALS['wgHooks']['TitleGetSquidURLs'][] = 'ForumHooksHelper::onTitleGetSquidURLs';
+$GLOBALS['wgHooks']['ArticleCommentGetSquidURLs'][] = 'ForumHooksHelper::onArticleCommentGetSquidURLs';
 
 // SUS-1196: Invalidate "Forum Activity" rail module when deleting a thread via Nuke / Quick Tools
-$wgHooks['ArticleDeleteComplete'][] = 'ForumHooksHelper::onArticleDeleteComplete';
+$GLOBALS['wgHooks']['ArticleDeleteComplete'][] = 'ForumHooksHelper::onArticleDeleteComplete';
 
 // SUS-260: Prevent moving pages within, into or out of Forum namespaces
-$wgHooks['MWNamespace:isMovable'][] = 'ForumHooksHelper::onNamespaceIsMovable';
+$GLOBALS['wgHooks']['MWNamespace:isMovable'][] = 'ForumHooksHelper::onNamespaceIsMovable';
 
-$wgHooks['AfterPageHeaderButtons'][] = 'ForumHooksHelper::onAfterPageHeaderButtons';
+$GLOBALS['wgHooks']['AfterPageHeaderButtons'][] = 'ForumHooksHelper::onAfterPageHeaderButtons';
 
 include ( $dir . '/Forum.namespace.setup.php' );