--- a/extensions/wikia/Forum/ForumActivityService.php
+++ b/extensions/wikia/Forum/ForumActivityService.php
@@ -1,0 +1,11 @@
+<?php
+
+interface ForumActivityService {
+	/**
+	 * Returns information about the N most recently modified Forum threads.
+	 * Information includes thread URL, thread title, author data, and timestamp.
+	 *
+	 * @return array
+	 */
+	public function getRecentlyUpdatedThreads(): array;
+}