--- a/extensions/wikia/Forum/templates/Forum_forumActivityModule.php
+++ b/extensions/wikia/Forum/templates/Forum_forumActivityModule.php
@@ -3,21 +3,21 @@
 	<ul class="activity-items">
 		<?php foreach ( $posts as $value ): ?>
 		<li class="activity-item">
-			<?php $username = User::isIP( $value['display_username'] ) ? wfMessage( 'oasis-anon-user' )->escaped() : htmlspecialchars( $value['display_username'] ); ?>
+			<?php $displayName = $value['authorId'] ? htmlspecialchars( $value['authorName'], ENT_QUOTES ) : wfMessage( 'oasis-anon-user' )->escaped(); ?>
 
-			<a class="activity-avatar" href="<?= $value['user']->getUserPage()->getFullUrl() ?>">
-				<img class="wds-avatar" src="<?= AvatarService::getAvatarUrl( $value['user']->getName(), 30 ) ?>" data-tracking="activity-avatar" title="<?= $username ?>" />
+			<a class="activity-avatar" href="<?= Sanitizer::encodeAttribute( $value['authorUrl'] ); ?>">
+				<img class="wds-avatar" src="<?= AvatarService::getAvatarUrl( $value['authorName'], 30 ) ?>" data-tracking="activity-avatar" title="<?= $displayName ?>" />
 			</a>
 
 			<div class="activity-info">
 				<div class="page-title">
-					<a href="<?= $value['wall_message']->getMessagePageUrl( true ); ?>" title="<?= Sanitizer::encodeAttribute( $value['metatitle'] ); ?>" data-tracking="activity-title"><?= htmlspecialchars( $value['metatitle'] ); ?></a>
+					<a href="<?= Sanitizer::encodeAttribute( $value['threadUrl'] ); ?>" title="<?= Sanitizer::encodeAttribute( $value['threadTitle'] ); ?>" data-tracking="activity-title"><?= htmlspecialchars( $value['threadTitle'] ); ?></a>
 				</div>
 
 				<div class="edit-info">
-					<a class="edit-info-user" href="<?= $value['user']->getUserPage()->getFullUrl() ?>" data-tracking="activity-username">
-						<?= $username ?>
-					</a><span class="edit-info-time"> • <?= wfTimeFormatAgo( $value['event_iso'] ) ?></span>
+					<a class="edit-info-user" href="<?= Sanitizer::encodeAttribute( $value['authorUrl'] ); ?>" data-tracking="activity-username">
+						<?= $displayName ?>
+					</a><span class="edit-info-time"> • <?= wfTimeFormatAgo( $value['timestamp'] ); ?></span>
 				</div>
 			</div>
 		</li>