… extensions/wikia/Forum/tests/fixtures/comments_index.sql
… extensions/wikia/Forum/tests/fixtures/recently_updated_threads.yaml