Optimize SQL query performance in Forum activity module.

Changes summary:
* extracted Forum specific code to Forum extension from Wall
* module now actually returns 5 entries when available, not some other number
* module now correctly detects if the thread was edited by the author, and reflects such updates
* module cache is no longer purged when a Wall thread (rather than a Forum one) is changed.

## Before the optimization
`WallHistory::getLastPosts` method would generate nested SQL query on cache miss:
```sql
SELECT `parent_page_id`,
       `post_user_id`,
       `post_user_ip_bin`,
       `is_reply`,
       `parent_comment_id`,
       `comment_id`,
       `action`,
       `event_date`,
       (SELECT `metatitle`
        FROM   `wall_history` AS `last_title`
        WHERE  `parent_comment_id` = `wall_history`.`parent_comment_id`
               AND ( `event_date`, `revision_id` ) = (SELECT Max(`event_date`),
                                                             Max(`revision_id`)
                                                      FROM   `wall_history`
                                                      WHERE
                   `action` IN ( 0, 1 )
                   AND `parent_comment_id` =
               `last_title`.`parent_comment_id`)
        LIMIT  1) AS `metatitle`,
       `reason`,
       `revision_id`
FROM   `wall_history`
       RIGHT JOIN(SELECT `parent_comment_id`,
                         Max(`event_date`)  AS `event_date`,
                         Max(`revision_id`) AS `revision_id`
                  FROM   `wall_history`
                         RIGHT JOIN(SELECT `parent_comment_id`
                                    FROM   `wall_history`
                                    WHERE  `action` = 1
                                           AND `deleted_or_removed` = 0
                                           AND `post_ns` = 2000
                                           AND `is_reply` = 0) AS
                                   `not_removed_parents`
                         USING (
                         `parent_comment_id`)
                  WHERE  `action` = 1
                         AND `deleted_or_removed` = 0
                         AND `post_ns` = 2000
                  GROUP  BY `parent_comment_id`
                  ORDER  BY `event_date` DESC
                  LIMIT  5) AS `last_new_post_date` USING (
       `parent_comment_id`, `event_date`, `revision_id`); 
```
`EXPLAIN` output:
```
+----+--------------------+--------------+------------+------+--------------------+--------------------+---------+--------------------------------------------------------------------+--------+----------+----------------------------------------------+
| id | select_type        | table        | partitions | type | possible_keys      | key                | key_len | ref                                                                | rows   | filtered | Extra                                        |
+----+--------------------+--------------+------------+------+--------------------+--------------------+---------+--------------------------------------------------------------------+--------+----------+----------------------------------------------+
|  1 | PRIMARY            | <derived4>   | NULL       | ALL  | NULL               | NULL               | NULL    | NULL                                                               |      2 |   100.00 | NULL                                         |
|  1 | PRIMARY            | wall_history | NULL       | ref  | getByParentComment | getByParentComment | 9       | last_new_post_date.parent_comment_id,last_new_post_date.event_date |      1 |   100.00 | Using where                                  |
|  4 | DERIVED            | wall_history | NULL       | ALL  | getByParentComment | NULL               | NULL    | NULL                                                               | 542702 |     0.01 | Using where; Using temporary; Using filesort |
|  4 | DERIVED            | wall_history | NULL       | ref  | getByParentComment | getByParentComment | 5       | wikia.wall_history.parent_comment_id                               |      2 |     1.69 | Using where                                  |
|  2 | DEPENDENT SUBQUERY | last_title   | NULL       | ref  | getByParentComment | getByParentComment | 5       | wikia.wall_history.parent_comment_id                               |      2 |   100.00 | Using where                                  |
|  3 | DEPENDENT SUBQUERY | wall_history | NULL       | ref  | getByParentComment | getByParentComment | 5       | wikia.last_title.parent_comment_id                                 |      2 |    20.00 | Using where                                  |
+----+--------------------+--------------+------------+------+--------------------+--------------------+---------+--------------------------------------------------------------------+--------+----------+----------------------------------------------+
6 rows in set, 3 warnings (0.00 sec)
```
Return values and execution time on `vsbattles` database:
```
+----------------+--------------+------------------+----------+-------------------+------------+--------+---------------------+------------------------------------------------------------+--------+-------------+
| parent_page_id | post_user_id | post_user_ip_bin | is_reply | parent_comment_id | comment_id | action | event_date          | metatitle                                                  | reason | revision_id |
+----------------+--------------+------------------+----------+-------------------+------------+--------+---------------------+------------------------------------------------------------+--------+-------------+
|          32586 |     32333285 | NULL             |        0 |           1376655 |    1376655 |      1 | 2018-02-12 16:15:24 | I narrowly dodge the Mami Spam (I think)                   | NULL   |     2190177 |
|          32586 |     28702329 | NULL             |        1 |           1373324 |    1376656 |      1 | 2018-02-12 16:15:24 | NOT EQUALIZED SPEED: Accelerator vs Doflamingo             | NULL   |     2190178 |
|          32586 |     30472761 | NULL             |        1 |           1373296 |    1376654 |      1 | 2018-02-12 16:15:09 | Hashirama Senju vs Sosuke Aizen                            | NULL   |     2190176 |
|           3523 |     26879614 | NULL             |        1 |            972086 |    1376653 |      1 | 2018-02-12 16:15:06 | Re:Zero Kara Hajimeru Isekai Seikatsu Discussion Thread 16 | NULL   |     2190175 |
|          32586 |     26984863 | NULL             |        1 |           1205906 |    1376651 |      1 | 2018-02-12 16:14:32 | Stabbing concepts vs Punching concepts                     | NULL   |     2190173 |
+----------------+--------------+------------------+----------+-------------------+------------+--------+---------------------+------------------------------------------------------------+--------+-------------+
5 rows in set (11.70 sec)
```

## After the optimization
A simpler query is used:
```sql
SELECT page.page_id                   AS page_id,
       page.page_namespace            AS page_namespace,
       page.page_title                AS page_title,
       thread_rev.rev_text_id         AS thread_text_id,
       thread_rev.rev_user            AS thread_author_id,
       thread_rev.rev_user_text       AS thread_author_name,
       thread_rev.rev_timestamp       AS thread_modified_timestamp,
       last_comment_rev.rev_user      AS last_comment_author_id,
       last_comment_rev.rev_user_text AS last_comment_author_name,
       last_comment_rev.rev_timestamp AS last_comment_timestamp
FROM   `page`
       INNER JOIN `comments_index`
               ON (( page.page_id = comment_id ))
       INNER JOIN `revision` `thread_rev`
               ON (( thread_rev.rev_id = page.page_latest ))
       INNER JOIN `page` `last_comment_page`
               ON (( last_comment_page.page_id = last_child_comment_id ))
       INNER JOIN `revision` `last_comment_rev`
               ON (( last_comment_rev.rev_id = last_comment_page.page_latest ))
WHERE  page.page_namespace = '2001'
       AND archived = '0'
       AND deleted = '0'
       AND removed = '0'
       AND parent_comment_id = '0'
       AND ( parent_page_id != 0 )
ORDER  BY last_touched DESC
LIMIT  5; 
```
`EXPLAIN` output:
```
+----+-------------+-------------------+------------+--------+-----------------------------------------------------+--------------+---------+------------------------------------------------+------+----------+-------------+
| id | select_type | table             | partitions | type   | possible_keys                                       | key          | key_len | ref                                            | rows | filtered | Extra       |
+----+-------------+-------------------+------------+--------+-----------------------------------------------------+--------------+---------+------------------------------------------------+------+----------+-------------+
|  1 | SIMPLE      | comments_index    | NULL       | index  | PRIMARY,comment_id,parent_comment_id,parent_page_id | last_touched | 16      | NULL                                           |   55 |     0.09 | Using where |
|  1 | SIMPLE      | page              | NULL       | eq_ref | PRIMARY,name_title,page_ns_latest_idx               | PRIMARY      | 4       | vsbattles.comments_index.comment_id            |    1 |    50.00 | Using where |
|  1 | SIMPLE      | last_comment_page | NULL       | eq_ref | PRIMARY                                             | PRIMARY      | 4       | vsbattles.comments_index.last_child_comment_id |    1 |   100.00 | NULL        |
|  1 | SIMPLE      | thread_rev        | NULL       | eq_ref | rev_id                                              | rev_id       | 4       | vsbattles.page.page_latest                     |    1 |   100.00 | NULL        |
|  1 | SIMPLE      | last_comment_rev  | NULL       | eq_ref | rev_id                                              | rev_id       | 4       | vsbattles.last_comment_page.page_latest        |    1 |   100.00 | NULL        |
+----+-------------+-------------------+------------+--------+-----------------------------------------------------+--------------+---------+------------------------------------------------+------+----------+-------------+
5 rows in set, 1 warning (0.01 sec)
```
Return values and execution time on `vsbattles` database:
```
+---------+----------------+------------------------------------------------------------+----------------+------------------+--------------------+---------------------------+------------------------+--------------------------+------------------------+
| page_id | page_namespace | page_title                                                 | thread_text_id | thread_author_id | thread_author_name | thread_modified_timestamp | last_comment_author_id | last_comment_author_name | last_comment_timestamp |
+---------+----------------+------------------------------------------------------------+----------------+------------------+--------------------+---------------------------+------------------------+--------------------------+------------------------+
| 1332021 |           2001 | General_Discussion/@comment-Veloxt1r0kore-20180130045757   |        2097255 |         31760473 |                    | 20180204053901            |               26865490 |                          | 20180212161338         |
| 1375402 |           2001 | Content_Revision/@comment-Thebluedash-20180212063148       |        2133568 |         26041742 |                    | 20180212063148            |               26984863 |                          | 20180212161256         |
| 1369182 |           2001 | Versus_Threads/@comment-The_real_cal_howard-20180210204333 |        2125118 |         27175658 |                    | 20180210204333            |               27175658 |                          | 20180212161252         |
| 1295448 |           2001 | Wiki_Management/@comment-Antvasima-20180118090303          |        2020863 |         24383490 |                    | 20180118090303            |               26474458 |                          | 20180212161239         |
| 1374761 |           2001 | Versus_Threads/@comment-Nico-v11-20180212023121            |        2134952 |         30742659 |                    | 20180212150021            |               26971556 |                          | 20180212161223         |
+---------+----------------+------------------------------------------------------------+----------------+------------------+--------------------+---------------------------+------------------------+--------------------------+------------------------+
5 rows in set (0.00 sec)
```

https://wikia-inc.atlassian.net/browse/SUS-4084